Update the related posts block
